+++ src/check/rules/mod.rs
@@ -8,12 +8,5 @@ mod clause;
 mod data;
 /// Check a list of declarations.
 mod decls;
-/// Synthesize the type and its well-typed form from an abstract term.
-mod infer;
 /// Type check a term.
 mod term;
-/// Conversion check.
-mod unify;
-/// Find the weak-head-normal-form (normalize) of an expression.
-/// TODO: Unfolds declarations.
-mod whnf;